requirements.txt
opencv
numpy
matplotlib
pillow
scikit-learn
scikit-image